Move config from nn to gui.legacy
