+++ sleap/gui/learning/receptivefield.py
@@ -7,7 +7,7 @@ import numpy as np
 from qtpy import QtWidgets, QtGui, QtCore
 
 from sleap import Video, Track, Skeleton
-from sleap.nn.config import ModelConfig
+from sleap.gui.legacy.config import ModelConfig
 from sleap.gui.widgets.video import GraphicsView
 from sleap.nn.model import Model
